moved remaining provisioners logic to core, added docstrings
