+++ molecule/core.py
@@ -230,3 +230,125 @@ class Molecule(object):
             return False
 
         return True
+
+    def _remove_templates(self):
+        """
+        Removes the templates created by molecule.
+
+        :return: None
+        """
+        os.remove(self._config.config['molecule']['vagrantfile_file'])
+        os.remove(self._config.config['molecule']['rakefile_file'])
+        os.remove(self._config.config['molecule']['config_file'])
+
+    def _create_templates(self):
+        """
+        Creates the templates used by molecule.
+
+        :return: None
+        """
+        # vagrantfile
+        kwargs = {
+            'config': self._config.config,
+            'current_platform': self._env['MOLECULE_PLATFORM'],
+            'current_provider': self._env['VAGRANT_DEFAULT_PROVIDER']
+        }
+        utilities.write_template(self._config.config['molecule']['vagrantfile_template'],
+                                 self._config.config['molecule']['vagrantfile_file'],
+                                 kwargs=kwargs)
+
+        # ansible.cfg
+        utilities.write_template(self._config.config['molecule']['ansible_config_template'],
+                                 self._config.config['molecule']['config_file'])
+
+        # rakefile
+        kwargs = {
+            'molecule_file': self._config.config['molecule']['molecule_file'],
+            'current_platform': self._env['MOLECULE_PLATFORM']
+        }
+        utilities.write_template(self._config.config['molecule']['rakefile_template'],
+                                 self._config.config['molecule']['rakefile_file'],
+                                 kwargs=kwargs)
+
+    def _create_inventory_file(self):
+        """
+        Creates the inventory file used by molecule and later passed to ansible-playbook
+
+        :return: None
+        """
+        inventory = ''
+        # TODO: for Ansiblev2, the following line must have s/ssh_//
+        host_template = \
+            '{} ansible_ssh_host={} ansible_ssh_port={} ansible_ssh_private_key_file={} ansible_ssh_user={}\n'
+        for instance in self._config.config['vagrant']['instances']:
+            ssh = self._vagrant.conf(vm_name=utilities.format_instance_name(instance['name'], self._env[
+                'MOLECULE_PLATFORM'], self._config.config['vagrant']['instances']))
+            inventory += host_template.format(ssh['Host'], ssh['HostName'], ssh['Port'], ssh['IdentityFile'],
+                                              ssh['User'])
+
+        # get a list of all groups and hosts in those groups
+        groups = {}
+        for instance in self._config.config['vagrant']['instances']:
+            if 'ansible_groups' in instance:
+                for group in instance['ansible_groups']:
+                    if group not in groups:
+                        groups[group] = []
+                    groups[group].append(instance['name'])
+
+        for group, instances in groups.iteritems():
+            inventory += '\n[{}]\n'.format(group)
+            for instance in instances:
+                inventory += '{}\n'.format(utilities.format_instance_name(instance, self._env['MOLECULE_PLATFORM'],
+                                                                          self._config.config['vagrant']['instances']))
+
+        inventory_file = self._config.config['molecule']['inventory_file']
+        try:
+            utilities.write_file(inventory_file, inventory)
+        except IOError:
+            print('{}WARNING: could not write inventory file {}{}'.format(Fore.YELLOW, inventory_file, Fore.RESET))
+
+    def _create_playbook_args(self):
+        """
+        Builds up CLI and ENV arguments from various config files to be passed to ansible-playbook.
+
+        :return: None
+        """
+        # don't pass these to molecule-playbook CLI
+        env_args = ['raw_ssh_args', 'host_key_checking', 'config_file', 'raw_env_vars']
+
+        # args that molecule-playbook doesn't accept as --arg=value
+        special_args = ['playbook', 'verbose']
+
+        # set raw environment variables if any are found
+        if 'raw_env_vars' in self._config.config['ansible']:
+            for key, value in self._config.config['ansible']['raw_env_vars'].iteritems():
+                self._env[key] = value
+
+        self._env['PYTHONUNBUFFERED'] = '1'
+        self._env['ANSIBLE_FORCE_COLOR'] = 'true'
+        self._env['ANSIBLE_HOST_KEY_CHECKING'] = str(self._config.config['ansible']['host_key_checking']).lower()
+        self._env['ANSIBLE_SSH_ARGS'] = ' '.join(self._config.config['ansible']['raw_ssh_args'])
+        self._env['ANSIBLE_CONFIG'] = self._config.config['ansible']['config_file']
+
+        kwargs = {}
+        args = []
+
+        # pull in values passed to molecule CLI
+        if '--tags' in self._args:
+            self._config.config['ansible']['tags'] = self._args['--tags']
+
+        # pass supported --arg=value args
+        for arg, value in self._config.config['ansible'].iteritems():
+            # don't pass False arguments to ansible-playbook
+            if value and arg not in (env_args + special_args):
+                kwargs[arg] = value
+
+        # verbose is weird -vvvv
+        if self._config.config['ansible']['verbose']:
+            args.append('-' + self._config.config['ansible']['verbose'])
+
+        kwargs['_env'] = self._env
+        kwargs['_out'] = utilities.print_stdout
+        kwargs['_err'] = utilities.print_stderr
+
+        return self._config.config['ansible']['playbook'], args, kwargs
